requirements.txt
yt-dlp>=2025.1.1
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
PyGithub>=2.3.0
